fix issues blocking deployment
