+++ packages/cloud/serverless.yml
@@ -50,5 +50,5 @@ functions:
     handler: refreshSpellIconsHandler
     events:
       - event:
-          eventType: google.cloud.pubsub.topic.v1.messagePublished
+          eventType: providers/cloud.pubsub/eventTypes/topic.publish
           resource: projects/${self:provider.project, ""}/topics/refresh-spell-icons-event
