+++ packages/cloud/src/writeMatchStubHandler.ts
@@ -2,49 +2,15 @@ import { Firestore } from '@google-cloud/firestore';
 import { instanceToPlain } from 'class-transformer';
 import fetch from 'node-fetch';
 
-import { IArenaMatch, IShuffleMatch, WoWCombatLogParser, WowVersion } from '../../parser/dist/index';
+import { WowVersion } from '../../parser/dist/index';
 import { createStubDTOFromArenaMatch, createStubDTOFromShuffleMatch } from './createMatchStub';
+import { parseFromStringArrayAsync } from './utils';
 
 const matchStubsFirestore = process.env.ENV_MATCH_STUBS_FIRESTORE;
 
 const firestore = new Firestore({
   ignoreUndefinedProperties: true,
 });
-
-type ParseResult = {
-  arenaMatches: IArenaMatch[];
-  shuffleMatches: IShuffleMatch[];
-};
-
-export function parseFromStringArrayAsync(
-  buffer: string[],
-  wowVersion: WowVersion,
-  timezone?: string,
-): Promise<ParseResult> {
-  return new Promise((resolve) => {
-    const logParser = new WoWCombatLogParser(wowVersion, timezone);
-
-    const results: ParseResult = {
-      arenaMatches: [],
-      shuffleMatches: [],
-    };
-
-    logParser.on('arena_match_ended', (data: IArenaMatch) => {
-      results.arenaMatches.push(data);
-    });
-
-    logParser.on('solo_shuffle_ended', (data: IShuffleMatch) => {
-      results.shuffleMatches.push(data);
-    });
-
-    for (const line of buffer) {
-      logParser.parseLine(line);
-    }
-    logParser.flush();
-
-    resolve(results);
-  });
-}
 
 // In the Google code they actually type file as `data:{}`
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
