+++ packages/cloud/test/test_file_stubify.ts
@@ -1,10 +1,10 @@
 import fetch from 'node-fetch';
 
 import { createStubDTOFromArenaMatch, createStubDTOFromShuffleMatch } from '../src/createMatchStub';
-import { parseFromStringArrayAsync } from '../src/writeMatchStubHandler';
+import { parseFromStringArrayAsync } from '../src/utils';
 
-const root_uri = 'https://storage.googleapis.com/wowarenalogs-public-dev-log-files-dev/WoWCombatLog-110422_233521.txt';
-fetch(root_uri).then((response) => {
+const rootUri = 'https://storage.googleapis.com/wowarenalogs-public-dev-log-files-dev/WoWCombatLog-110422_233521.txt';
+fetch(rootUri).then((response) => {
   console.log('x-goog-meta-ownerid', response.headers.get('x-goog-meta-ownerid'));
   response.text().then(async (textBuffer) => {
     const parseResults = await parseFromStringArrayAsync(textBuffer.split('\n'), 'retail');
